pull example 2 code to HEAD, with a bit of spit and polish again

+++ cloudnative-appconfig/cloudnative-connections/src/main/java/com/corneliadavis/cloudnative/connections/ConnectionsController.java
@@ -11,12 +11,16 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import org.slf4j.Logger;
 
+@RefreshScope
 @RestController
 public class ConnectionsController {
 
     private static final Logger logger = LoggerFactory.getLogger(ConnectionsController.class);
     private UserRepository userRepository;
     private ConnectionRepository connectionRepository;
+
+    @Value("${com.corneliadavis.cloudnative.connections.secret}")
+    private String configuredSecret;
 
     @Autowired
     Utils utils;
@@ -28,20 +32,30 @@ public class ConnectionsController {
     }
 
 	@RequestMapping(method = RequestMethod.GET, value="/users")
-	public Iterable<User> getUsers(HttpServletResponse response) {
+	public Iterable<User> getUsers(@RequestParam(value="secret", required=true) String secret,
+                                   HttpServletResponse response) {
 
+        if (secret.equals(configuredSecret)) {
 
             logger.info(utils.ipTag() + "getting users");
             Iterable<User> users;
             users = userRepository.findAll();
             return users;
+        } else {
+            logger.info(utils.ipTag() + "Attempt to access Connections service with secret " + secret + " (expecting " + configuredSecret + ")");
+            response.setStatus(401);
+            return null;
+        }
     }
 
 	@RequestMapping(method = RequestMethod.GET, value="/users/{user}")
 	public User getByUsername(@PathVariable("user") String user,
+                              @RequestParam(value="secret", required=true) String secret,
                               HttpServletResponse response) {
 
-            logger.info(utils.ipTag() + "Accessing posts");
+        if (secret.equals(configuredSecret)) {
+
+            logger.info(utils.ipTag() + "Accessing posts using secret " + secret);
             String ipAddress = System.getenv("POD_IP");
             logger.info(utils.ipTag() + "getting user " + user);
             try {
@@ -50,69 +64,118 @@ public class ConnectionsController {
             } catch (NumberFormatException e) {
                 return userRepository.findByUsername(user);
             }
+        } else {
+            logger.info(utils.ipTag() + "Attempt to access Connections service with secret " + secret + " (expecting " + configuredSecret + ")");
+            response.setStatus(401);
+            return null;
+        }
     }
 
     @RequestMapping(method = RequestMethod.POST, value = "/users")
     public void newUser(@RequestBody User newUser,
+                        @RequestParam(value = "secret", required = true) String secret,
                         HttpServletResponse response) {
+
+        if (secret.equals(configuredSecret)) {
 
             logger.info(utils.ipTag() + "Have a new user with username " + newUser.getUsername());
             userRepository.save(newUser);
 
+        } else {
+            logger.info(utils.ipTag() + "Attempt to access Connections service with secret " + secret + " (expecting " + configuredSecret + ")");
+            response.setStatus(401);
+        }
     }
 
     @RequestMapping(method = RequestMethod.PUT, value = "/users/{id}")
     public void updateUser(@PathVariable("id") Long userId,
                            @RequestBody User newUser,
+                           @RequestParam(value = "secret", required = true) String secret,
                            HttpServletResponse response) {
+
+        if (secret.equals(configuredSecret)) {
 
             logger.info(utils.ipTag() + "Updating user with id " + userId);
             User user = userRepository.findOne(userId);
             newUser.setId(userId);
             userRepository.save(newUser);
 
+        } else {
+            logger.info(utils.ipTag() + "Attempt to access Connections service with secret " + secret + " (expecting " + configuredSecret + ")");
+            response.setStatus(401);
+        }
     }
 
     @RequestMapping(method = RequestMethod.GET, value = "/connections")
-    public Iterable<Connection> getConnections(HttpServletResponse response) {
+    public Iterable<Connection> getConnections(@RequestParam(value = "secret", required = true) String secret,
+                                               HttpServletResponse response) {
+
+        if (secret.equals(configuredSecret)) {
 
             logger.info(utils.ipTag() + "getting connections");
             Iterable<Connection> connections;
             connections = connectionRepository.findAll();
 
             return connections;
+        } else {
+            logger.info(utils.ipTag() + "Attempt to access Connections service with secret " + secret + " (expecting " + configuredSecret + ")");
+            response.setStatus(401);
+            return null;
+        }
     }
 
     @RequestMapping(method = RequestMethod.GET, value = "/connections/{username}")
     public Iterable<Connection> getConnections(@PathVariable("username") String username,
+                                               @RequestParam(value = "secret", required = true) String secret,
                                                HttpServletResponse response) {
+        if (secret.equals(configuredSecret)) {
 
             logger.info(utils.ipTag() + "getting connections for username " + username);
-            Long userId = getByUsername(username, null).getId();
+            Long userId = getByUsername(username, secret, null).getId();
             Iterable<Connection> connections;
             connections = connectionRepository.findByFollower(userId);
 
             return connections;
+        } else {
+            logger.info(utils.ipTag() + "Attempt to access Connections service with secret " + secret + " (expecting " + configuredSecret + ")");
+            response.setStatus(401);
+            return null;
+        }
     }
 
     @RequestMapping(method = RequestMethod.POST, value = "/connections")
     public void newConnection(@RequestBody Connection newConnection,
+                              @RequestParam(value = "secret", required = true) String secret,
                               HttpServletResponse response) {
+
+        if (secret.equals(configuredSecret)) {
 
             logger.info(utils.ipTag() + "Have a new connection: " + newConnection.getFollower() + " is following " + newConnection.getFollowed());
             connectionRepository.save(newConnection);
 
+        } else {
+            logger.info(utils.ipTag() + "Attempt to access Connections service with secret " + secret + " (expecting " + configuredSecret + ")");
+            response.setStatus(401);
+        }
     }
 
     @RequestMapping(method = RequestMethod.DELETE, value = "/connections/{id}")
     public void deleteConnection(@PathVariable("id") Long connectionId,
+                                 @RequestParam(value = "secret", required = true) String secret,
                                  HttpServletResponse response) {
+
+        if (secret.equals(configuredSecret)) {
 
             Connection connection = connectionRepository.findOne(connectionId);
 
             logger.info(utils.ipTag() + "deleting connection: " + connection.getFollower() + " is no longer following " + connection.getFollowed());
             connectionRepository.delete(connectionId);
 
+        } else {
+            logger.info(utils.ipTag() + "Attempt to access Connections service with secret " + secret + " (expecting " + configuredSecret + ")");
+            response.setStatus(401);
+        }
     }
+
 
 }
